MNIST/crl/requirements.txt
torchvision
tqdm
matplotlib
numpy
git+https://github.com//ncullen93/torchsample#0.1.3
pandas
nibabel
